Add use of maven
