+++ .gitignore
@@ -2,3 +2,6 @@ build.bat
 *.log
 *.class
 src/*.class
+.mvn/
+.idea/
+target/
