--- lib/javafx.properties
@@ -1,3 +1,0 @@
-javafx.version=23.0.1
-javafx.runtime.version=23.0.1+4
-javafx.runtime.build=4
