--- src/AlgorithmGenerator.java
@@ -1,8 +1,0 @@
-package src;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public interface AlgorithmGenerator {
-    AlgorithmStepList generate(ListUI visualList, HighlightableCodeArea codeArea, ArrayList<Integer> integers, List<String> args);
-}
