+++ src/main/java/edu/carleton/algorithm/AlgorithmArgParser.java
@@ -1,0 +1,15 @@
+package edu.carleton.algorithm;
+
+public class AlgorithmArgParser {
+    public interface SortPred {
+        boolean comp(int x, int y);
+    }
+
+    public static SortPred parseSortPred(String angleBracketStr) {
+        if (angleBracketStr.equals("descending")) {
+            return (x, y) -> { return x > y; };
+        } else {
+            return (x, y) -> { return x < y; };
+        }
+    }
+}
