+++ src/main/java/edu/carleton/algorithm/AlgorithmGenerator.java
@@ -1,0 +1,8 @@
+package edu.carleton.algorithm;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public interface AlgorithmGenerator {
+    AlgorithmStepList generate(ListUI visualList, HighlightableCodeArea codeArea, ArrayList<Integer> integers, List<String> args);
+}
